Fix the `MethodDescriptor.specializeTypeVariables`.
The modeling of method descriptors is quite nuanced. Both declaration and paramterized references are modeled with the same class `MethodDescriptor`. Declarations always have empty type arguments even when they have type parameters. `MethodDescriptor.specializeTypeVariables` needs to take this into account, since computing the specialized type arguments is different whether the descriptor represents a declaration or a parameterized reference.

This cl also makes sure that the invariant that declarations have empty type arguments is preserved.

There are no usages yet that rely on the type arguments that are the result of specialization, so in practice this is a noop.

PiperOrigin-RevId: 613978311