Automated g4 rollback of commit 44f23ae020ee2b9df27dcd909783229c3e2b0b56.
*** Reason for rollback ***

Break Kix-j2kt-web

*** Original change description ***

Generate the same code for Interface's companion instance field than for classes' companion.

For interfaces, Kotlin/JVM generates a private static field containing the unique instance of the companion and a public static field on the interface exposing the companion private field. This is not needed in our case and removing this construct will help our future optimization of companions.

***

PiperOrigin-RevId: 622914711