Fix import collection for Overlay types.
During the import collection for a overlay type, we collect types that needs to be imported from the underlying native type. During this process we only need to collect types that are part of the native api. Everything that comes from the overlay api will be collected from the overlay type.

PiperOrigin-RevId: 626244835