Fix our logic to determine if a field is compile time constant in Kotlin.
Our current logic tests if the initializer of a field is a literal constant to mark a field as compile time constant.  In Kotlin, `const val` IrField initialized to their default value does not have an initializer expression and is not marked as compiletime constant in the J2CL ast.

PiperOrigin-RevId: 627088890