[KT] Move backing field of properties defined on the interface companion to the enclosing interfaces.
Due to the JVM restrictions on private interface members, Kotlin/JVM does not move backing field of properties defined on companion interfaces except if they are all annotated with `JvmField`.
In Closure, we do not have that restrictions and we can safely move all backing fields to the enclosing interface. That brings the way we process interface companion properties in parity with class companion properties.

PiperOrigin-RevId: 628210752