[KT] Fix incorrect handling of supercalls in companion objects.
Due to kotlin lowerings done on the IR super method calls might not have a `super` as their qualifier. That is the case for companions where the lowering replaces `super` and `this` references by references to the singleton companion field. In those cases, the method calls need to be marked as static dispatch in the J2CL AST.

PiperOrigin-RevId: 631882170