[KT] Remap type parameters from fake override functions to the resolved ones
Currently we're not remapping them so our type parameter map will be referring
to the type variables from the fake override, despite the function
resolved to something else.

The type parameters should effectively be the same (both in order and name), but
for good measure I remap them based on the indices.

This should resolve issues where we couldn't specialize types on method calls
as the type parameters didn't match what was in the map.

PiperOrigin-RevId: 632550033