Optimize `ForEachStatement`s over well-known Xplat iterable types
There's two scenarios that we'll handle:

  1. array-like types which will become
     `for (int = 0; i < arr.length; i++) { var element = arr[i]; ... }`
  2. object-like types will become `for (let property in obj) { ... }`

So long as we encounter a `ForEachStatement` where we can unambiguously identify
these usages, we'll rewrite them. There are a couple edge cases where we need
to be mindful of string/int coercion and auto-unboxing implicitly occuring in
the variable assignment.

Additionally, to support this I introduced a new `JsForInStatement` that it
similar to `ForEachStatement`, but is intended to encapsulate the JS `for..in`
semantics. This is needed since we can have both `JsForInStatement` and
`ForEachStatement` in the AST at the same time and thus need a way to keep track
of the different semantics.

PiperOrigin-RevId: 633263389