[J2KT] Disable `getExtension/hasExtension` rewrite
The additional import requires an extra build dependency that we’d technically
have to synthesize in our build macros. Kotlin Native doesn’t enforce strict
dependencies yet, so this was only a problem on JVM. The original names don’t
have that problem.

Also, for the new protos, it was a design goal to implement a subset of the java lite proto API, the `com.google.protobuf.kotlin` APIs are not part of that. So in preparation for the new protos, we’re switching to the Java-style names (old protos have also been updated to support them).

PiperOrigin-RevId: 634286022