[J2KT] Filter protobuf-related APIs from J2ObjC compat headers
The protos are not generated by J2KT but direcly generated as Kotlin, so they
don’t follow J2KT J2ObjC-compatible naming conventions. This was not a problem
with old protos (which were value classes, so exported to ObjC as `id`) but new
protos are exported and lead to invalid compat headers. Skip generating
compatibility APIs for anything involving protos for now.

PiperOrigin-RevId: 636530585