[J2KT] Add an example of incorrect overload resolution with absent varargs
Java and Kotlin have different overload resolution semantics when invoking a
member with absent varargs. Java resolves to a member that doesn't take varargs,
whereas Kotlin resolves to a member that accepts varargs, even though they're
absent in the call.

Currently J2KT doesn't resolve this ambiguity, which can be solved by adding a
cast to force the resolution to a particular member.

PiperOrigin-RevId: 638850421