[J2KT] Add casts to arguments where the target method is ambiguous
In particular, if there are two candidate call, for example:

```
void foo(Object o) { ... }
void foo(String o, Object... rest) { ... }
```

with a caller of `foo("bar")`, Java will resolve this calling the first method.

Currently this is faithfully brought over to Kotlin as `foo("bar")`. However,
Kotlin has different semantics for overload resolution, preferring the most
specific candidate. In that case it will end up invoked `foo(String, Object...)`
which is inconsistent with the original code.

To workaround this we'll detect cases where there is a compatible overload that
differs only on variable arity, and add casts to the specific parameter types
of the intended target method.

PiperOrigin-RevId: 640903632