Use a different native type in importglobaljstypes
The `NativeFunction` class is overlaying `Function`, but it's declaring a type
parameter that doesn't actually exist on the underlying type. This doesn't have
any real effect in the current readable since we never render any code for this.

However, when this code is ported to Kotlin there will be a bridge method on the
companion object that would actually result in the `NativeFunction<String>`
being rendered as `Function<String>`, which will result in a type error.

To avoid this I just rewrote the code to use `IArrayLike<T>`. I also added a
method that returns a `IArrayLike<T>` so that we actually generate a reference
to the type. This would have expose the type error in the original code.

While I was here I also renamed the method `fun` to `f` to avoid using Kotlin
reserved keyword.

PiperOrigin-RevId: 647002378