[KT] Don't extract a tmp variable for try-statements used as expressions
The existing logic create a tmp variable of the try-statement result type and
then assign the result of the try and all catch blocks to that tmp variable.

However, if the try isn't being used as an expression then the result type of
the try and catch blocks do not need to be assignable to one another. In which
case the try will have `Unit` result type, which we'll attempt to assign the
block results to. This will ultimately produce type errors in JS.

Instead we'll back-off if the try-statement type is `Unit`, opting to just
create a composite statement that explicitly return `Unit`.

PiperOrigin-RevId: 648913784