Simplify `transitivejsoverlayimport` readable
This mostly involved creating real instances rather than just assigning to null,
as this otherwise requires null checks in Kotlin that don't really make sense.
I additionally needed to make the JsProperty reference a real property that
actually exists on `Object` to get the JSC to not complain (surprisingly it
allowed the non-existent property reference on `null`).

I also removed the extra blocks as they're not providing any value to this test.

PiperOrigin-RevId: 649147387