Honor JsProperty annotation for fields that don't generate accessors
Private properties in Kotlin elide the accessors entirely and just use the
backing field directly. Normally we don't allow backing fields to participate
in JsInterop, but here we should since there's no accessors to apply the
`@JsProperty` semantics to.

Unfortunately this is complicated to figure out in the Kotlin IR as the
`IrProperty` will still have a getter/setter associated with it, even if it's
the default. The trick is that these default functions don't end up in the
declarations of the class, which is why we don't generate code for them today.

This made the logic somewhat complex as I need to both check that the property
is private and that it lacks a user-defined (i.e. non-default) getter or setter.

PiperOrigin-RevId: 650389584