[J2KT] Add a warning when transpiling a type that is not `@NullMarked`
In general J2KT only produces correct results when the input code is both
`@NullMarked` and nullability tested.

Currently if you give it code that is not `@NullMarked`, you may just end up
with incorrect output code without any insight into what went wrong. Ideally
this would be an error, but there's a lot of pre-existing usages that would have
to be burned down first.

I did back-off on annotation types since those are generally not a problem, and
it's fairly common for annotation-specific packages to lack `@NullMarked`.

PiperOrigin-RevId: 655960428