Make `OptimizeAnonymousInnweClassesToFunctionExpression` replace the type of the anonymous class with functional interface.
The original pass assumed that the type was not leaked because it was non denotable. However with the introduction of `var` in Java and the Kotlin frontend, variables can be inferred to be of that type.

This cl fixes the immediate problem but does not replace types in method signatures, which might also experience the same issue.

PiperOrigin-RevId: 657203318