[KT] Implement enumEntries intrinsic
This can be invoked in two ways: `kotlin.enums.enumEntries<SomeEnum>()` or
`SomeEnum.entries`. The former is only available in Kotlin 2.0 whereas the
latter is available in 1.9.

Normally for the JVM backend the `entries` property getter is implemented by
compiler and `enumEntries()` is rewritten at call sites. In both cases it reads
the cached entries from a static field on the enum, which is initialized by
simply calling `enumEntries(SomeEnum.values())`.

For J2CL this gets complicated since we don't generate the `values()` method for
enums that opted into `@JsEnum`, so I do not want to generate an `entries`
getter on all enums. To avoid this I've opted to rewrite all usages of both
call variants at the call site. This will replace them directly with
`enumEntries(SomeEnum.values())`, which if it's a `@JsEnum`, will fail
JsInterop restrictions. It will be a bit of a confusing message since the user
didn't actually write a call to `values()`, but I think it's good enough for
now.

In the future we can improve this by actually implementing `entries` and add
restriction checks to it. We further can optimize it by lazily caching the
result.

PiperOrigin-RevId: 660843224