Determine the target for an override signature in a more principled manner.
The resolution of which method is the target for an override signature in abstract classes was done in a bit of an ad-hoc manner. This cl implements the resolution in a principled way relying upon the invariants of correct programs.

The target of an override signature is a method for that signature that will be implementor of that specific signature (all other methods with the same override signature will bridge to it). In many cases this is easy to determine, e.g. if there is a non-abstract method in the class hierachy, that would be the target.

The general algorithm determines the target by the following cases:

  - if there is a method defined in the class hierachy, that is the target. In this case, all other methods with the same override signature are either overridden by it and their return type is not specialized further.
  - if there is no method in the class hierarcy, find the method a method that overrides all the candidates, if one is found that is the target. Note that if the target is implemented by a default method all the other methods having the same override signature need to be explicilty overridden by the default method.
  - if there are still multiple candidates, they will all be abstract (all default methods explicitly overridden by abstract methods) and the one with the more specific return type will be the target.

PiperOrigin-RevId: 660851431