Don't perform full IR serialization for libraries using J2KT
For JVM builds we will only set `-Xserialize-ir` if it's a non-J2KT target, otherwise we don't set the flag at all.

For J2CL builds we will always set the `-Xserialize-ir` flag as this also enables deserialization of the IR. We do not actually produce any serialized IR in these builds.

PiperOrigin-RevId: 662184988