[JS/WASM] Support non-native JsEnum arrays.
In Closure, make new arrays use Arrays.createNative.
We don't need metadata for JsEnum arrays because we don't perform type checking at runtime. Simple changes in `NormalizeArrayCreations`. We store the unboxed value for JsEnums in Closure.

In Wasm, `JsEnum`s are considered normal objects (`WasmArray.OfObject`). No special changes are needed aside from restrictions changes. `JsEnum`s are transpiled like regular enums, which are also stored as array of object.

Restrictions:
We disallow JsEnum arrays to be assigned to Object[], T[], T..., and so on. The following fails: `Object[] o = new MyJsEnum[0];` and users cannot check 'instanceof MyJsEnum[]`. See `JsInteropRestrictionsCheckerTest` for the new cases.
We want to allow JsEnum arrays, but restrict it for now to hide the runtime type and avoid boxing/unboxing scenarios and implementation logic.

JsEnum ARRAYS are not allowed to be passed to varargs, but individual JsEnums are allowed in varargs. See `ErasePackagedJsEnumVarargs`.

This CL does not affect native arrays of JsEnum. The pass `OptimizeXplatForEach` correctly considers native arrays of JsEnum to contain the boxed type.

PiperOrigin-RevId: 665467270