[KT] Replace K2 incompatible API for resolving type.
`pluginContext.referenceClass(FqName)` is no longer supported in K2. In preparation to switching to K2, this cl replaces call to this function with call to `pluginContext.referenceClass(ClassId)` that is supported both by the old and the new frontend.

Standardize all the ClassId creations by always using the java binary qualified name string format.

PiperOrigin-RevId: 665958383