[WASM] Enable JsEnums.
In preparation for implementing JsEnums in Wasm protobuf, enable non-native JsEnums in Wasm by setting the flag in `_impl_j2wasm_application` and `_j2cl_transpile` (for modular)

`@JsEnum` annotations for non-native JsEnums will not be removed on creation of the `TypeDeclaration`. Native JsEnums still have the annotation removed.

Note: This also enables support for JsEnums in modular. For Wasm, JsEnum is currently just an optimization and does not mean they are accessible from JS.
PiperOrigin-RevId: 665995332