Use ASM based package-info resolver in JDT and Kotlin frontends.
We were already using a variation of this based on loading class w/ ClassLoader, then we refactored it out of our code to use JDT based resolution due to problems with that. However JDT is not great at one of type resolutions and causes quite bit of performance hit (seen up to 30% of compilation cost).

As of yesterday, new version version PackageInfoCache uses ASM to load the class so we can go back to it.

I sent out of couple cleanup CLs. I will also follow up this CL with a change that removes singleton around PackageInfoCache to make things slightly better. I tried couple of other failed refactoring and I'm still not super happy with the existing setup.

We need to think more around JdtEnvironment's mock initialization and how PackageAnnotationResolver relates with other utilities and gets instantiated but will stick with existing setup for now.

PiperOrigin-RevId: 668981056