[WASM] Explicitly throw exceptions from Wasm with throw instruction.
In the past there was a V8 bug that when we throw a JSTAG exception, it was still converted to a WebAssembly exception and the original JS error was lost. This has been fixed a while back (over a year ago) so we don't need to call JS to throw the exception.

PiperOrigin-RevId: 675427342