Account for the kotlin rename of `T List<T>.remove(int)` to `removeAt` in override related computations.
Kotlin renames this method to avoid clashing `List<T>.remove(T)` when specializing to `List<Int>` and since the signature of a method (used as key in the JVM) includes the return type, it can still emit class files that are interoperable with Java by not changing the name in the signature.

In this cl we take the same approach and only use the kotlin name for considerations about overridding methods (e.g. to compute bridges, etc).

PiperOrigin-RevId: 684478225