Use a canonical representation of the enclosing type descriptor in method declarations.
Since the identity of methods is determined by the identity of declarations, declarations need to be build consistently. In particular the nullability of the enclosing type was left to the usage site leading to having multiple declarations for the same method only differing in the nullaibility of the enclosing type.

PiperOrigin-RevId: 684824069