Simplify code around creation of DeclaredTypeDescriptor.
This CL adds toDescriptor overload to TypeDeclaration that allows you to create a DeclaredTypeDescriptor with a specific set of type argument descriptors. This method could be used as the main building block for almost all DeclaredTypeDescriptor construction.

The rest of constructions could be replace with "DeclaredTypeDescriptor.withTypeArguments" but didn't go so far to keep the CL smaller.

PiperOrigin-RevId: 686355035