Rework the backing machinery for System.getProperty to utilize our own registry
The existing system is brittle properties are read as global symbols when
uncompiled, but as `goog.define` flags when compiled. This is error-prone and
there's no clear linkage between the two.

To replace it, system properties now need to be explicitly registered. For now
we'll only be supporting registering properties that come from a `goog.define`
flag.

For backwards compatibility we'll fallback to `goog.getObjectByName` if a
property is not registered, but we will make this an error in the future.

PiperOrigin-RevId: 689188398