Add an IrProvider that loads IR nodes based on their public signature using the IrPluginContext.
This provider is essential for IR deserialization with the K2 compiler. During deserialization, unbound symbols are generated for references to external declarations that haven't been loaded yet. In the K1 compiler, an IrProvider relying on the descriptor API is used to load or create IR nodes for these symbols. However, the descriptor API is no longer available in K2. This provider addresses this by utilizing the public signature of the unbound symbol and the IR plugin API to resolve the corresponding IR node.

PiperOrigin-RevId: 690797578