Make sure that passes that are run before flattening don't visit contents of types more than once.
Using `NormalizationPass.applyTo(Type)` is error prone before the type tree is flattened and should be avoided.

Since types will appear inside the type AST, some operations become error prone or not intuitive.

For example visiting the methods of a type:

```
applyTo(Type type) {
  type.accept( new AbstractVisitor() {
    void exitMethod(Method method) {
      ...
    }
  }
}
```

In that case `exitMethod` will not only visit the methods of `type` but also the methods of the nested types. And since `applyType`  will be called for every type in the tree, methods will be vistied a number of times equal to their nesting depth.

PiperOrigin-RevId: 693065932