[KT] Add ExternalPackageParentPatcherLowering .
The K2 frontend does not create a FileKt class  for referenced top-level members defined in another compilation unit but attaches them to an IrPackageFragment. This lowering addresses this by creating the necessary file class and moving the top-level member into it. This is needed for:
  - our CUB that expects any members to be defined in an enclosing class
  - IR deserialization of top level inline members (The serialized IR being part of the metadata of the enclosing class after Kotlin/JVM compilation.)

This is a copy of the same lowering pass from Kotlinc.

This is a noop if K2 is not enabled.

PiperOrigin-RevId: 694182541