Remove enforcing the use of the "sources" classifier from the maven_coordinates rule.
This is a no-op for the Guava case (only current example in the repo) since its GWT artifact actually includes the sources so we didn't need forcing a "sources" classifier to begin with.

For the Truth (which I'm playing with and triggered this patch), it doesn't have a separate artifact but correct download is chosen with a classifier. The change here will make sure the user of the rule can choose the correct target by a precise maven coordinates (e.g. com.google.truth:truth:jar:gwt:1.4.4 where option "jar:gwt" in middle defines the packaging and classifier).

PiperOrigin-RevId: 696222517