Move JvmStaticInObjectLowering to run after IntrinsicFunctionCallsLowering
This is important as intrinsics may be lowered to function calls on object
instances. While this currently works, it's only because we see these functions
as being static. In 2.1.0 they are corrected and we become reliant upon
JvmStaticInObjectLowering to rewrite all the callers. Therefore, it must
occur after all intrinsics have been substituted.

PiperOrigin-RevId: 700106997