Preserve @JsAsync annotations through J2KT
Unfortunately I cannot enable the integration test for this as we don't have a
way to write the special `await` method that will come through J2KT and
Kotlin/Closure nicely. The problem is that the static method will be brought
over as a `@JvmStatic` companion function. However, Kotlin/Closure will move the
"real" function to the enclosing class, and the companion member will serve as
a bridge method.

That's a problem though since calls to `await()` will be rewritten to just the
`await` keyword. This won't parse since you cannot use `await` in a non-`async`
function.

We'll need to rework how the integration test is structured in order to get this
to work.

PiperOrigin-RevId: 700804967