Ensure we remap the types parameters on both the proxy and static functions
When you have a `@JvmStatic` external member, you will end up with a pair of
newly created functions. One will be moving the external member to the enclosing
class, and the second will be a new proxy function on the companion object.
Currently we are are only remapping the type parameters on one of these
functions.

This does result in some readable changes since we're not properly specializing
method descriptors, which causes us to emit erasure casts.

PiperOrigin-RevId: 702060952