Pass through extra JS files generated by APT through J2KT
I needed to do stuff the extra JS outputs into a zip file in order for it to flow through J2CL and maintain the relative output path. If I just output into a directory that I create, then output post-J2CL will include both the path to the directory and the relative path within that directory.

If we don't care about those paths then we would be able to simplify this a bit.

PiperOrigin-RevId: 703216273