[KT] Resolve primitive types annotated with `@EnhancedNullability` as boxed types
When JSpecify is being strictly enforced Kotlin will actually honor the
JSpecify nullability it sees in Java when constructing types in its frontend. When it does this it tags the type with an `@EnhancedNullability` annotation.

This is problematic though if we are need to interop with NullMarked Java code
that refers to a non-nullable boxed types. Since Kotlin doesn't represent
boxed/unboxed (only null/non-null), we end up a non-nullable Kotlin primitive
type. Normally we would assume this to be unboxed, but with the nullability
enhancement it's actually supposed to be boxed.

To resolve this we'll assume any primitive type annotated with `@EnhancedNullability`
is actually a non-nullable boxed type, which matches how Kotlin/JVM does
[signature mapping](https://cs.android.com/android-studio/kotlin/+/master:core/compiler.common.jvm/src/org/jetbrains/kotlin/load/kotlin/typeSignatureMapping.kt;drc=059046a2b1dc9ff352ebe09d222ddc1dca6272ef;l=41-45).

This did result in a change to an existing readable test. I'm not entirely sure
where the enhanced nullability is coming from in this example though, but I did verify that the adjusted signatures now match those in the header jar.

PiperOrigin-RevId: 706222231