[J2KT] Remove `JavaCollection/JavaList/JavaSet` bridge types
On JVM/J2CL, these are now replaced with typealiases of the plain Kotlin `MutableCollection/MutableList/MutableSet` interfaces.

On Kotlin Native, they are replaced with the emulated `java.util.Collection/List/Set` interfaces which are sub-interfaces of the Kotlin interfaces.

The `toArray` methods are not visible on the interfaces in JVM/J2CL so we need to suppress warnings regarding the `override` modifier (but only for direct implementations of the interface methods. The much more common case of overriding AbstractCollection#toArray does not have this requirement).

As a result, binaries (.class/.js/.klib) will all use java.util.Collection (etc.) as the implemented/extended interface.

PiperOrigin-RevId: 712572687