Force well-known Kotlin MethodDescriptors to be marked static
These functions are marked with `@JvmStatic` so there will be a static bridge
method present, but starting in Kotlin 2.1, the frontend will no longer mark
these as static. Instead it will treat them as a instance access on the
singleton object. It then relies upon lowering passes to rewrite the code.

However, this is problematic for these well-known descriptors as the Kotlin
frontend will now tell us the method isn't static, whereas a Java frontend will
tell us that it is.

Since these are a restricted set of well-known methods, we can trust that we've
appropriately marked them `@JvmStatic` and force the method descriptor to be
marked static.

PiperOrigin-RevId: 713340500