Remove flag that disables timestamp reseting in output files.
Instead always reset time for zip outputs (where it is needed) and never for directory output.

Since we are using directory outputs with Bazel this seems like spending around 500ms-1000s (3-5%) for some large targets. However since we offload I/O work to another thread and try to parallelize it, it is hard to tell actualized improvement (plus I/O performance is generally higher on Forge).

PiperOrigin-RevId: 713796112