Upgrade Kotlin/Closure to Kotlin 2.1.0
For now the standard library is being pinned to the previous version as there's
quite a few changes that need to be made as it's finally upgrading to K2.

This is largely all just compiler API changes, with a few tweaks to IR
deserialization provided by dramaix@.

One concerning change is that function inlining got quite a bit more verbose
with additional temporary variables being created. We'll likely need to follow
up on that as the code is now quite complicated to read and may inhibit the JSC
from optimizing as well as it otherwise could.

PiperOrigin-RevId: 714104373