Use a dedicated heap type for the JS Tag
Binaryen started preserving the specific heap types used by  tags rather than generating new heap types for them based on their signatures when writing binaries. As a result, the imported JS tag started using whatever function type happened to be the first to be declared to take a single externref parameter and return no results. In general, this type is not the expected type when importing the JS exception tag. To fix the issue, declare and use the precise expected function type in the WAT output.

PiperOrigin-RevId: 717026300