[JS] Improve handling of precedence for comma expressions.
Allows for rendering parenthesis in components of multiexpressions if their precedence is lower than that of the COMMA expression. Still multiexpressions as a whole are rendered with enclosing parenthesis and considered of the highest precedence. Ideally we could model all of this correctly with the proper precedence but jscompiler has inconsistent handling of some constructs.

PiperOrigin-RevId: 718047008