[KT] For functions coming from Java, use the special bridge types, if applicable
There are cases where Kotlinc presents us false types for things coming from
Java, collections in particular. For example, the `contains` function on
Collections will always be considered as having a `contains(T)` signature,
despite the Java actually having a `contains(Any)` signature. This is
problematic if the Java specialized the type (for example with `String`) as it
would present as `contains(String)`.

When we bring these incorrect types into the J2CL AST is can throw things off
as we'll think something exists that doesn't. For example, a Kotlin subtype
overriding `contains` will now thing it's overriding `contains(String)` when
that method signature does not actually exist in the parent type.

To resolve this I've stolen some of the bridge construction logic to substitute
param and return types for these functions.

PiperOrigin-RevId: 721978389