[KT] Generate bridge methods where Kotlin and Java names don't align
We currently only generate Kotlin bridges for special methods, but don't
generate them for generic bridges or bridges to correct name mismatches. We skip
the generic bridges as those can be generated in the J2CL AST but we do miss the
the name mismatches.

Currently the names mismatches are "corrected" by mapping the name. That is, if
the Kotlin name is `getSize` and the Java name is `size`, we just substitute
`getSize` for `size`. However, this is problematic if Java then subtypes the
Kotlin as it will think that the `getSize` method does exist.

Kotlin typically resolves these mismatches by bridging the Java super-method
to delegate to the Kotlin variant of the method.

I opted to remove much of the logic that we have for resolving function names,
delegating much of the logic back to the Kotlin/JVM backend. The only "trick" I
needed to do here is compute that pretending that functions are not internal
to avoid applying the name mangling. There are additional complications involved
when mangling names so I'm deferring to b/236236685 for that.

I also needed to apply a correction to the bridge for `remove`/`removeAt` to
avoid erasing the return type. This isn't a complete solution but it brings us
back in line to the existing behavior. Improvements for that specific mapping
are being tracked in b/372484266.

PiperOrigin-RevId: 726185977