Add cancelation support for Kotlin frontend.
This currently needs to be wired in via ProgressIndicatorAndCompilationCanceledStatus. Noticeably they are missing a lot of checks in between works that causes many large 1000ms+ delays in my local testing. If it turns out to be a problem, we will need to change couple of places to add checks in more places within kotlinc.

Also unfortunately the mechanism they have relies on static state so doesn't support multiplex workers. In the future I might workaround that using thread locals but for now I kept it simple.

PiperOrigin-RevId: 728739574