Fix cancelation of Kotlin frontend.
It looks like throwing our own exception causes Kotlin to wrap with its own and propagate like a failure. CL changes the mechanism to throw CompilationCanceledException which is properly handled by kotlinc and so it gracefully exits.

We didn't catch this earlier since existing tests are only testing the delay and prevent any exception to be thrown. I have some tests in the working so this might be partially covered by tests (though it is very difficult to fully repro this kind of problems in the integration tests).

PiperOrigin-RevId: 729227621