[KT] Convert to != directly from != and !==
Currently Kotlin will always represent `!=` and `!==` as `!(==)` and `!(===)` in
their AST, which we'll faithfully bring over as `!(==)` (or `!(.equals())`).

Instead we can look at the origin for not binary operator as well as what we
already converted the operand to. If we know that not originally came from `!=`
or `!==` and the operand has already been converted to an infix `==`, we can
simply rewrite the infix to be `!=`.

This should hopefully make the code a bit easier to read and look much closer
to what the original code was.

PiperOrigin-RevId: 735789178