Avoid depending on Kotlin targets for package-info compilation.
This creates a new java compilation at Bazel level and does best effort(1) removal of kotlin dependencies for the new target.
In most cases, these separate dependencies won't change hence so will be cached well and also would be a much smaller subset to compile against.
On the other side, it will create an extra step (<0.2s) in cold builds but currently this a better trade-off.

Since I moved the compilation outside, I also removed package-info parsing code from the Kotlin frontend.

(') This is not bullet-proof due to j2cl_library#exports but good enough for common cases for the compilation avoidance and I wanted to keep it simple.

PiperOrigin-RevId: 736960497