Throw an error if addSystemPropertyFromGoogDefine is called after getSystemProperty
This is intended to guard cases where someone attempts to read a property before
it's registered. This would lead to diverging behavior between compiled and
uncompiled code as the former would return the value computed statically at
compile time whereas the latter would return undefined.

To help developers catch this mistake we'll now throw an error during
registration of where the property was first attempted to be read.

PiperOrigin-RevId: 740011131