[KT] Persist binary assignment operators from Kotlin
When we encounter an assignment we can check the origin of the statement to
figure out what the original assignment operator was. If both the LHS and RHS
are primitives then we can directly bring this over to the J2CL AST.

Note: this does impact int and friends as these get normalized in the J2CL AST
to ensure that we have proper bounds on them.
PiperOrigin-RevId: 741288732