Smuggle several JVM lowering passes to expose them to J2CL
Since we generally only need access to the ctors of lowering passes, we can
abuse Kotlin/Java interop semantics (i.e. internal becomes public) to "smuggle"
these types into the Kotlin frontend, avoiding our need to fork the code.

The only trickery that needs to occur is that the `KotlinInternal` check needs
to be suppressed.

PiperOrigin-RevId: 743777553