Add target based profiling for js/wasm compilation actions.
Following a similar approach to Kotlin rules, profiled targets are configured via build setting //third_party/java_src/j2cl:profiling_filter. Profiling works for both workers and remote execution.

The solution pays attention to keep the compiler binary intact when there is no profiling (avoids any java and launcher/swig deps). Also it is implemented in a way that it could be integrated to any of our rules that is built on top of BazelWorker with couple of lines (though for now I only integrated with js/wasm transpile actions).

PiperOrigin-RevId: 744857758