Pass javaemul annotations to `-classpath` for Wasm modular transpilation and the Wasm bundler
BootClasspathInfo has a `DepSet` `_auxiliary` field. This change uses this field to pass dependencies of the JRE.

This is to allow the transpiler to read annotation values for `javaemul.internal` annotations, which are not part of the classpath unless one of the dependencies explicitly includes it as a dependency.  We were currently working around this by using the internal `compiler.lookup` package for JDT because these annotations could not be accessed via DOM.

PiperOrigin-RevId: 746670096