Roll forward of https://github.com/google/j2cl/commit/33d43a4278a0132719e7a102ede6849806c4d103: [J2KT] Don't synthesize jl.Object methods on AutoValue types
By not synthesizing these methods we don't run afoul of the expectations in the
OptimizeAutoValue pass. If these methods are present on the AutoValue-annotated
class then AutoValue will not generate them on the subtype. OptimizeAutoValue
relies upon this expectation and will fail if these methods are present on both
the AutoValue class and the generated subtype.

AutoValue classes are always abstract and will always override these methods
from jl.Object, without calling back to the super method, so not synthesizing
the methods is safe. If this were to ever change, the Kotlin code would fail to
compile.

It's important to note that the AutoValue APT runs on the original Java code
before J2KT runs. So AutoValue does not see these methods being defined on the
AutoValue class and thus generates them on the subtype. However, J2KT then runs
and added the bridges to the AutoValue class. This is ultimately how we ended up
in an unexpected state.

PiperOrigin-RevId: 751162114