Allow lateinit property initialization checks to be disabled
Currently kotlinc inlines null check, which if the field is null, calls the
`throwUninitializedPropertyAccessException()` intrinsic. This doesn't give us
much flexibility so I've opted to hard fork the `LateinitLowering` to allow us
to fully change how we generate code for these.

Separately, the `JvmPropertiesLowering` is currently responsible for inlining
private lateinit usages as the JVM backend doesn't generate getters/setters for
private properties. To avoid also having to fork this lowering pass as well
I've instead incorporated the logic for getters into our lowering passes.

With this change lateinit initialization checks will now be disabled when API
checking is disabled.

PiperOrigin-RevId: 755021996