[KT] Collect fields from Java classes that are now represented as Kotlin properties.
Since Kotlin 2.0.20, fields from Java class are loaded as kotlin properties. We need to unwrap the backing field of these property in order to correctly construct our type model.

This cannot be directly tested because any reference to these filed from Kotlin are correctly targeting the backing field. The problem happens when we try to read the field from the type descriptor in the J2CL backend. This will be indirectly tested through our coroutine implementation that need such logic.

PiperOrigin-RevId: 758360397