[KT] Fix inlining of top-level functions called between different files in the same module
Fixes a Kotlin inliner bug where a top-level inline function, defined in one file, was not inlined when called from another inline function in a different file within the same or in another module.

The failure occurred because the parent of the target inline function was not correctly patched after its call site (from the other file's inline function) was inlined. This incorrect patching prevented the deserialization of the target function's IR, thereby blocking the inlining.

In This CL, I copied over the visitor used in ExternalPackageParentPatcherLowering instead of forking that pass and augment it's visibility. This would have introduced a cyclic dependency.

PiperOrigin-RevId: 761227404